Adding a new book to the manager’s array of books works, but it is not displayed in the tableview.

+++ Phase 2/Book.h
@@ -12,7 +12,7 @@
 
 @property NSString *author;
 @property NSString *title;
-@property int price;
+@property NSInteger price;
 @property NSString *isbn;
 @property NSString *course;
